Added Cross Promotion Template
Created a Cross Promotion Template for Laurence Krulik. This is very
similar to the Famous Person Landing Page, but instead  of hardcoding
the image to Taylor Swift, it accepts a URL to an image. This will be
changed to selecting a banner, once that functionality has been fixed on
the site.